Tests/CSM.BLTest/ProductRepositoryTest.cs
using System;
using CMS.BL;
using Microsoft.VisualStudio.TestTools.UnitTesting;

namespace CSM.BLTest
{
    [TestClass]
    public class ProductRepositoryTest
    {
        [TestMethod]
        public void RetrieveTest()
        {
            //--Arrange
            var productRepository = new ProductRepository();
            var expected = new Product(2)
            {
                CurrentPrice = 10.5M,
                ProductDescription = "Point of sale application",
                ProductName = "R10"
            };
            //--Act
            var actual = productRepository.Retrieve(2);

            //--Assert
            Assert.AreEqual(expected.CurrentPrice, actual.CurrentPrice);
            Assert.AreEqual(expected.ProductDescription, actual.ProductDescription);
            Assert.AreEqual(expected.ProductName, actual.ProductName);
        }
    }
}
